+++ src/openrct2/object/SceneryGroupObject.cpp
@@ -83,7 +83,7 @@ void SceneryGroupObject::UpdateEntryIndexes()
             continue;
 
         uint16_t sceneryEntry = objectManager.GetLoadedObjectEntryIndex(ori->LoadedObject);
-        Guard::Assert(sceneryEntry != UINT8_MAX, GUARD_LINE);
+        Guard::Assert(sceneryEntry != UINT16_MAX, GUARD_LINE);
 
         auto objectType = ori->ObjectEntry.flags & 0x0F;
         switch (objectType)
